refactor: seal client exceptions
