+++ packages/serverpod_client/lib/serverpod_client.dart
@@ -4,5 +4,4 @@ export 'src/auth_key_provider.dart';
 export 'src/connectivity_monitor.dart';
 export 'src/serverpod_client_exception.dart';
 export 'src/serverpod_client_shared.dart';
-export 'src/method_stream/method_stream_manager_exceptions.dart';
 export 'src/file_uploader.dart';
