+++ packages/serverpod_client/lib/src/method_stream/method_stream_manager_exceptions.dart
@@ -1,4 +1,4 @@
-import 'package:serverpod_client/serverpod_client.dart';
+part of '../serverpod_client_exception.dart';
 
 /// Exceptions thrown by the [ClientMethodStreamManager].
 abstract class MethodStreamException implements Exception {
